Inherit CST, AST, and SymbolTable classes from Tree class.
