+++ lib/marvin.rb
@@ -4,6 +4,7 @@ require_relative 'marvin/runner'
 require_relative 'marvin/command'
 require_relative 'marvin/logger'
 require_relative 'marvin/error'
+require_relative 'marvin/tree'
 
 require_relative 'marvin/grammar'
 require_relative 'marvin/token'
