+++ lib/marvin/node.rb
@@ -4,7 +4,7 @@ require 'securerandom'
 module Marvin
 
   # Subclass and namespace +Tree::TreeNode+.
-  class Node < Tree::TreeNode
+  class Node < ::Tree::TreeNode
 
     def initialize(content)
       super(SecureRandom.hex, content)
